move test data factory to root level
